+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Custom request for client to get comments generated for a given line. This logic simply finds first instance of keyword from a line and create a comment (regular or doxygen style) depending on the type of keyword.
+
+### Fixed
+
+- Returning of undefined node from `getNodeFromNameArray` function is fixed to return a valid value - when found.
+
 ## [0.3.1] - 2025-04-10
 
 ### Fixed